Moved contract tests to contracts subdirectory
